Confirm Booking Page

+++ src/app/password/password.page.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { ToastController } from '@ionic/angular';
+import { AlertController, ToastController } from '@ionic/angular';
 import { TranslateService } from '@ngx-translate/core';
 import { PassengerService } from '../services/passenger.service';
 
@@ -15,6 +15,7 @@ export class PasswordPage implements OnInit {
     public passengerService: PassengerService,
     public toastController: ToastController,
     public t: TranslateService,
+    public alertControler: AlertController,
     public r: Router,
   ) {
     t.get("passwordPage").subscribe((resp: any) => {
@@ -34,12 +35,24 @@ export class PasswordPage implements OnInit {
     email: '',
     password: ''
   }
+  async AlreadyLoggedIn() {
+    const alert = await this.alertControler.create({
+      header: this.respFromLanguage.opps,
+      message: this.respFromLanguage.alreadyLogin,
+      buttons: ['OK']
+    });
 
+    await alert.present();
+  }
   next() {
     this.passengerService.login(this.data).subscribe((resp: any) => {
-      localStorage.setItem('user', JSON.stringify(resp.pessenger));
-      localStorage.removeItem('TempUser');
-      this.r.navigate(['/home'])
+      if (resp.message == 'You are already login from another device. Please logput first to login from this device') {
+        this.AlreadyLoggedIn();
+      } else {
+        localStorage.setItem('user', JSON.stringify(resp.pessenger));
+        localStorage.removeItem('TempUser');
+        this.r.navigate(['/home'])
+      }
     }, err => {
       this.presentToast(this.respFromLanguage.inCorrect)
     })
